Semana5/02-django/entornovirtual/requirements.txt
asgiref==3.3.1
click==7.1.2
Django==3.1.4
djangorestframework==3.12.2
pip-tools==5.4.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
